Introducing NamespaceSelector as own component
